Overhaul CorrectionPipelineRunner. Move CorrectionPipelineRunner into correct package.
Fix InvokeFlash and InvokeQuake so that they work with the local job runner.
InvokeFlash uses ShellUtil to execute external commands so that we catch errors.

+++ src/main/java/contrail/correct/CutOffCalculation.java
@@ -23,9 +23,12 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.StringTokenizer;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.mapred.JobConf;
 import org.apache.hadoop.mapred.RunningJob;
 import org.apache.hadoop.util.ToolRunner;
 import org.apache.log4j.Logger;
@@ -35,33 +38,34 @@ import contrail.stages.Stage;
 import contrail.util.FileHelper;
 
 /**
- *  Cutoff calculation helps us in determining which kmers are trusted and untrusted. 
+ *  Cutoff calculation helps us in determining which kmers are trusted and untrusted.
  *  Trusted Kmers are kmers that have a frequency of occurrence more than a threshold
  *  (cutoff) value in the input dataset. They are used for correction, because
  *  their presence is more likely than the presence of kmer with a low frequency
  *  count (untrusted kmer). Therefore, trusted kmers are used for correction purposes
  *  by the quake engine.
- *  cov_model.py uses VGAM package within R to run some statistical analysis to 
+ *  cov_model.py uses VGAM package within R to run some statistical analysis to
  *  calculate the cutoff value. All kmers below the cutoff are considered untrusted.
- *  This is a non MR job. 
+ *  This is a non MR job.
  *  For calculating the cutoff, we need a fragment of the Kmer count file in text format
  *  A small portion of the file (fragment) is enough for calculating the cutoff.
  *  A fragment of the kmer count file is copied from the HDFS onto the local system where cov_model.py is
- *  run on it to calculate the cutoff value. The value is read from the output stream of the 
+ *  run on it to calculate the cutoff value. The value is read from the output stream of the
  *  cov_model.py process.
  */
-
 public class CutOffCalculation extends Stage {
   private int cutoff;
   private static final Logger sLogger = Logger.getLogger(CutOffCalculation.class);
- 
-  public int getCutoff() throws Exception{
-    if(cutoff!=0){
-      return cutoff;
+
+  public int getCutoff() {
+    if (cutoff == 0) {
+      sLogger.fatal("Cutoff wasn't calculated",
+          new Exception("ERROR: Cutoff not calculated"));
+      System.exit(-1);
     }
-	else throw new Exception("ERROR: Cutoff not calculated");
+    return cutoff;
   }
-	
+
   /**
    * we only need a sample of all KMers counts to compute the cutoff value
    * This method calculates the cutoff by:
@@ -71,28 +75,73 @@ public class CutOffCalculation extends Stage {
    */
   public void calculateCutoff() throws Exception{
     //inputPath is the path of the file on DFS where the non avro count part is stored
-    String inputPath = (String) stage_options.get("inputpath");
+    Path inputPath = new Path((String) stage_options.get("inputpath"));
     String covModelPath = (String) stage_options.get("cov_model");
-    String tempWritableFolder = FileHelper.createLocalTempDir().getAbsolutePath();
-    File countFile = new File(tempWritableFolder, "kmerCountFile");
-    if (countFile.exists()){
-      countFile.delete();
+
+    // Check if inputPath is a directory.
+    if (inputPath.getFileSystem(getConf()).getFileStatus(inputPath).isDir()) {
+      // Construct a glob path to match all part files.
+      inputPath = new Path(FilenameUtils.concat(
+          inputPath.toString(), "part-?????"));
+      sLogger.info(
+          "inputpath is a directory. The following glob will be used to " +
+          "locate the input. Glob: " + inputPath.toString());
     }
-    Configuration conf = new Configuration();
-    FileSystem fs = FileSystem.get(conf);
-    Path hdfsInputPath = new Path(inputPath);
-    Path localCountPath = new Path(countFile.getAbsolutePath());
-    fs.copyToLocalFile(hdfsInputPath, localCountPath);
+
+    FileStatus[] matchedFiles =
+        inputPath.getFileSystem(getConf()).globStatus(inputPath);
+
+    if (matchedFiles.length != 1) {
+      sLogger.fatal(String.format(
+          "More than 1 file matched the input glob %s. Number matched:%d",
+          matchedFiles.length), new IllegalArgumentException());
+      System.exit(-1);
+    }
+    inputPath = matchedFiles[0].getPath();
+    sLogger.info("Using input:" + inputPath);
+    // Check if the input is already on the local filesystem and if not copy
+    // it.
+    // TODO(jeremy@lewi.us): What's the best way to check if its the local
+    // filesystem?
+    String countFile = null;
+    String tempWritableFolder = null;
+    if (!inputPath.getFileSystem(getConf()).getUri().getScheme().equals(
+            "file")) {
+
+      // TODO(jeremy@lewi.us): We should cleanup this temporary directory
+      // after we are done.
+      tempWritableFolder =
+          FileHelper.createLocalTempDir().getAbsolutePath();
+
+      FileSystem fs = FileSystem.get(getConf());
+      Path localCountPath = new Path(countFile);
+      sLogger.info(String.format("Copy %s to %s", inputPath, localCountPath));
+      fs.copyToLocalFile(inputPath, localCountPath);
+    } else {
+      countFile = inputPath.toUri().getPath();
+    }
+
     // command to run cov_model.py
-    String command = covModelPath+" --int "+ countFile.getAbsolutePath();
+    String command = covModelPath+" --int "+ countFile;
     cutoff = executeCovModel(command);
-    File tempFile = new File(tempWritableFolder);
-    if(tempFile.exists()){
-      FileUtils.deleteDirectory(tempFile);
+
+    // Clean up the temporary directory if we created one.
+    if (tempWritableFolder != null) {
+      File tempFile = new File(tempWritableFolder);
+      if(tempFile.exists()){
+        FileUtils.deleteDirectory(tempFile);
+      }
     }
   }
-	
-  private int executeCovModel(String command) throws Exception{
+
+  private int executeCovModel(String command) throws Exception {
+    // TODO(jeremy@lewi.us): This is very brittle we aren't detecting whether
+    // the cutoff calculation succeeded or failed. We should use the functions
+    // in ShellUtil. We should redirect the output
+    // Furthermore, the python script is firing off an R script so R is
+    // required.
+    // It might be easier if we just called the R script directly ourselves
+    // rather than using cov_model.py
     StringTokenizer tokenizer;
     String line;
     int calculatedCutoff = 0;
@@ -104,7 +153,7 @@ public class CutOffCalculation extends Stage {
       tokenizer = new StringTokenizer(line);
       /* Everything displayed by the execution of cov_model.py here is stored in tokenizer
        * line by line. In the end, the token containing
-       * the cutoff is taken out 
+       * the cutoff is taken out
        */
       if(tokenizer.hasMoreTokens() && tokenizer.nextToken().trim().equals("Cutoff:")){
         String ss = tokenizer.nextToken();
@@ -115,12 +164,12 @@ public class CutOffCalculation extends Stage {
    p.waitFor();
    return calculatedCutoff;
   }
-  
+
   protected Map<String, ParameterDefinition> createParameterDefinitions() {
     HashMap<String, ParameterDefinition> defs =new HashMap<String, ParameterDefinition>();
     defs.putAll(super.createParameterDefinitions());
     ParameterDefinition QuakeHome = new ParameterDefinition(
-    "cov_model", "location of cov_model.py", String.class, new String(""));
+        "cov_model", "location of cov_model.py", String.class, null);
     for (ParameterDefinition def: new ParameterDefinition[] {QuakeHome}) {
       defs.put(def.getName(), def);
     }
@@ -131,12 +180,25 @@ public class CutOffCalculation extends Stage {
   }
 
   public RunningJob runJob(){
+    // Check for missing arguments.
+    String[] required_args = {"cov_model"};
+    checkHasParametersOrDie(required_args);
+    logParameters();
+    Configuration base_conf = getConf();
+    JobConf conf = null;
+    if (base_conf != null) {
+      conf = new JobConf(getConf(), this.getClass());
+    } else {
+      conf = new JobConf(this.getClass());
+    }
+    setConf(conf);
     try{
-    calculateCutoff();
-    sLogger.info("Cutoff: " + getCutoff());
+      calculateCutoff();
+      sLogger.info("Cutoff: " + getCutoff());
     }
     catch(Exception e){
-      sLogger.error(e.getStackTrace());
+      sLogger.fatal("Failed to compute cutoff.", e);
+      System.exit(-1);
     }
     return null;
   }
